Added rtkit_realtime_simple() to rtkit.[ch].
This is a turnkey one-function-call solution to getting realtime scheduling with rtkit.
It is designed to make it easier to get realtime scheduling.

There are many details to getting realtime scheduling working that the user doesn't need to know about.
For example, rtkit_make_realtime fails if RLIMIT_RTTIME is too high, and most people are fine with using the maximum priority.
And many people won't be using dbus anywhere else in their code.
These people don't need to know how to get a dbus system bus connection.
All of these details are taken care of in rtkit_realtime_simple().
The error handling is deliberately simple because many people will not care why rtkit_realtime_simple() fails, and will just adopt some kind of less-performant fallback, or simply abort if the program must be run in realtime.